refactor(refactor): move logic to main.ts
- move most logic to main.ts
- don't need to call run in internal tests
